requirements.txt
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
flask==3.0.0
flask-cors==4.0.0
python-telegram-bot==20.7
pandas==2.1.3
numpy==1.26.2
websocket-client==1.6.4
python-dateutil==2.8.2